Added methods to board class
