+++ lib/player.rb
@@ -1,0 +1,10 @@
+# frozen_string_literal: true
+
+class Player
+  attr_accessor :name, :symbol
+
+  def initialize(name, symbol)
+    @name = name
+    @symbol = symbol
+  end
+end
